+++ history.md
@@ -1,6 +1,7 @@
 # 版本信息
 ## 2.16.0 (2018-01-26)
 * [ADD] `yyl` 新增 `--logLevel` 参数
+* [ADD] `yyl` `config` 新增 `config.proxy.ignores` 参数
 * [EDIT] 简化 构建工具 `optimize` 时界面
 * [EDIT] 分离 `yyl watch` 中 `server` 与 `optimize` 到 两个独立的窗口
 * [EDIT] `yyl` 构建方式改成内联，不再通过 `util.runCMD(gulp watch)` 的方式进行调用